+++ packages/syft/src/syft/core/tensor/lazy_repeat_array.py
@@ -112,7 +112,7 @@ class lazyrepeatarray:
     def sum(self, *args: Tuple[Any, ...], **kwargs: Any) -> np.ndarray:
         if "axis" in kwargs and kwargs["axis"] is None:
             # TODO: make fast
-            return self.to_numpy().sum(dtype=np.int32)
+            return self.to_numpy().sum()
         else:
             raise Exception("not sure how to do this yet")
